+++ src/lib/berlekamp.c
