+++ src/lib/crcgen.c
