+++ src/lib/ecc.h
