+++ src/lib/galois.c
