+++ src/lib/rs.c
